draw-and-tell/backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
python-multipart
boto3
tensorflow-macos==2.14.0
protobuf==4.25.1
torch
transformers
Pillow
pydantic
soundfile
pydub
ffmpeg
datasets